+++ src/main/java/com/dmm/task/controller/LoginController.java
@@ -1,4 +1,4 @@
-package com.dmm.task;
+package com.dmm.task.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,7 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class LoginController {
 	@GetMapping("/loginForm")
-	String loginForm() {
+	
+	String loginForm() {		
 		return "login";
 	}
 }
