+++ src/main/java/com/dmm/task/data/repository/TaskRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dmm.task.data.entity.Posts;
 
-public interface PostsRepository extends JpaRepository<Posts, Integer> {
+public interface TaskRepository extends JpaRepository<Posts, Integer> {
 
 }
