+++ src/main/java/com/dmm/task/form/TaskForm.java
@@ -5,7 +5,7 @@ import javax.validation.constraints.Size;
 import lombok.Data;
 
 @Data
-public class PostForm {
+public class TaskForm {
 	// titleへのバリデーション設定を追加
 	@Size(min = 1, max = 200)
 	private String title;
