+++ src/main/resources/templates/main.html
@@ -66,12 +66,12 @@
       <p class="calendar_title_text sd">SAT</p>
      </div>
     </div>
-    <div class="calendar_row sd" th:each="week : ${matrix}">
+    <div class="calendar_row sd" th:each="week : ${Month}">
      <div class="calendar_box  sd" th:each="day : ${week}">
       <div>
        <a th:href="@{/main/create/{date}(date=${day})}">[[${day.dayOfMonth}]]</a>
       </div>
-      <div th:each="t : ${tasks.get(day)}">
+      <div th:each="t : ${calendar.get(day)}">
        <a th:if="${t}" th:href="@{/main/edit/{id}(id=${t.id})}"> <span
         th:if="${t.done}">✅</span><span th:text="${t.title}"></span></a>
       </div>
